Validation and Excecption commit

+++ src/main/java/com/example/getmesocialservice/resource/UserResource.java
@@ -1,13 +1,18 @@
 package com.example.getmesocialservice.resource;
 
+import com.example.getmesocialservice.exception.RestrictedIfoException;
 import com.example.getmesocialservice.model.User;
 import com.example.getmesocialservice.service.AlbumService;
 import com.example.getmesocialservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import javax.validation.Valid;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Resource
 @RestController
@@ -16,7 +21,11 @@ public class UserResource {
     @Autowired
     private UserService userService;
     @PostMapping
-    public User saveUser(@RequestBody User user){
+    public User saveUser(@RequestBody @Valid User user) throws RestrictedIfoException {
+        if(user.getName().equalsIgnoreCase("root"))
+        {
+            throw new RestrictedIfoException();
+        }
         return userService.saveUser(user);
     }
     @GetMapping()
@@ -25,15 +34,27 @@ public class UserResource {
         return userService.getAllUser();
     }
     @PutMapping()
-    public User updateUserById(@RequestBody User user){
+    public User updateUserById(@RequestBody @Valid User user) throws RestrictedIfoException {
+        if(user.getName().equalsIgnoreCase("root"))
+        {
+            throw new RestrictedIfoException();
+        }
         return userService. updateUserById(user);
     }
     @DeleteMapping
-    public void deleteUser(@RequestParam( name = "userId") String userId){
+    public void deleteUser(@RequestParam( name = "userId") String userId) throws RestrictedIfoException {
+        if(userId.equalsIgnoreCase("root"))
+        {
+            throw new RestrictedIfoException();
+        }
         userService.deleteUser(userId);
     }
     @GetMapping("/find")
-    public List<User> getUserByName(@RequestParam(name ="name") String name)
-    {return userService. getUserByName(name);
+    public List<User> getUserByName(@RequestParam(name ="name") String name) throws RestrictedIfoException {  if(name.equalsIgnoreCase("root"))
+    {
+        throw new RestrictedIfoException();
     }
+        return userService. getUserByName(name);
+    }
+
 }
